Implement latch to lock
Now when a latched key is pressed it becomes locked. Press it again to
release it.
The status code has been redesigned with a FSM table for latch to lock
to function. I believe it is easier to read and maintain.
Note: the xrecord back-feed no longer works. This needs to be fixed in a
later commit.