Work in progress: add ramble mode support
This is the first commit to implement ramble mode.
Ramble mode will make use of pointer gestures as input. This is a
more efficient way to input text without clicking, as supposed.
This commit draws the mouse path that will be interpreted later on as
gestures. It currently works with the -with-ramble configure option. A
gconf option will be added in a later commit.
Still to do:
* Interpret the path as gesture and act on it.
* Add a setting option.
* Add proper credit to the initial submiter of the idea and
first experimental implementation.