Add the "over" placement for extensions
The new "over" placement allow to place a keyboard extension over the
rest of the keyboard.
This is not very useful as of yet but it will be useful when the action
keys are implemented.
The action keys will allow to activate and deactivate an extension.
The goal is to provide something like an "123" key that makes the
numeric keyboard appear over.