Add input method configuration option
3 input methods are now selectable from the configuration dialog.
* button: the default input method. Keys are pressed with the mouse
button.
* timer: the old auto-click mode. Keys are pressed when hovered for long
enough. The timeout is configurable.
* ramble: the new ramble mode. Keys are pressed when the distance of the
pointer movement is long enough over the key.

Note: the ramble mode is still experimental and is not compiled by
default. Use the --with-ramble configure option to compile it.
The ramble algorithm is not suitable for all use cases. More algorithms
will be added. See the Ardesia project for another algorithm that will
be integrated in the future.