Add support for at-spi2 (at-spi over dbus)
This is still experimental support.
The GNOME projects is deprecating at-spi over CORBA so this move is necessary.
The configure script checks for at-spi2 and falls back to at-spi over corba if not present.
New compilation dependancy: libatspi (dev package).
Running dependancy: at-spi2-registryd (from the at-spi2-core package)
The at-spi2 atk bridge is needed for the keyboard to work with gtk.