Work around gtk bug 556006
Fix yet another regression.
A key may stay focused when the pointer leaves the window under some
extreme conditions that are rare enough to not be ever experienced.
This workaround is not perfect but it is better to have it because
without it, the keyboard would act funny on Ubuntu Natty and probably on
Oeniric too.